rename: util 함수 분리
